Send emails using Python.
Hi, I am Tarun, hope you all are having a good time.
This is a small effort to making lives better with python without even having a deep knowledge of this beautiful language. 
Using the line of codes provided in the file you can do thousand of emails in a single day with just one click. 
You don't have to be a pro at Python, as I am also a beginner looking to find my way into the programming world.  

I am the one who got inspired by RealPython's article so thanks a lot, if you find my code hard you can shift to the source as a whole description is provided in the article.
https://realpython.com/python-send-email/